fix: add missing tables to drop fix: change columns to proper data types

+++ backend/src/db/sql/schema/003_create_weapon_tables.sql
@@ -84,7 +84,7 @@ CREATE TABLE IF NOT EXISTS weapon_stats
     flux_per_shot numeric,
     emp integer,
     impact integer,
-    turn_rate integer,
+    turn_rate numeric,
     op_cost integer,
     chargeup numeric,
     chargedown numeric,
